Pomodoro * Fixed Pomodoro not showing in SysTray * Fixed SysTray loader * Moved SystemTrayManager back to Managers namespace * Cleaned up VeraCrypt add-in manifest errors for when we enable it
